requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
Django==4.2.11
django-crispy-forms==1.14.0
djangorestframework==3.15.0
djangorestframework-api-key==2.2.0
gunicorn==20.1.0
html5lib==1.1
mechanize==0.4.9
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.2
typing_extensions==4.10.0
webencodings==0.5.1
